Add files via upload
This is the landing page of a hospital. In this landing page ,it includes a menu bar consists of home , about , facilities , our doctors, contact, on clicking on them it directs to the particular position present in the landing page. For this i have used some web images as well as few images present in my device. 
HOPE YOU ALL LIKE THIS.